robot/requirements_llm.txt
# LLM Robot Control Dependencies
openai>=1.0.0
fastapi
uvicorn
python-multipart
python-dotenv

# Voice Control (TTS)
elevenlabs
